update curl to curl-7.53.1.
